requirements.txt
Flask
requests
python-dotenv
openai==0.28.0